Move `constrainValue` and `spinValue` back to main index file
